updated utils to have a function per file. Updated permissions config to remove lists, and allow different levels of read
