+++ lib/utils/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  traverse: require('./traverse'),
+  allowed: require('./allowed')
+};
